Move pkg/config/seccomp package to internal/config/seccomp
We do not want to expose these implementations to en-users
so we make them private.
